requirements/requirements.txt
pytz==2020.5  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==8.1.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat
flower==0.9.7  # https://github.com/mher/flower
uvicorn[standard]==0.13.3  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.0.11  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.10.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4  # https://github.com/django-compressor/django-compressor
django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.12.2  # https://github.com/encode/django-rest-framework
django-cors-headers==3.7.0 # https://github.com/adamchainz/django-cors-headers

Werkzeug==1.0.1 # https://github.com/pallets/werkzeug
ipdb==0.13.4  # https://github.com/gotcha/ipdb
psycopg2==2.8.6  # https://github.com/psycopg/psycopg2
watchgod==0.6  # https://github.com/samuelcolvin/watchgod

# Testing
# ------------------------------------------------------------------------------
mypy==0.800  # https://github.com/python/mypy
django-stubs==1.7.0  # https://github.com/typeddjango/django-stubs
pytest==6.2.2  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.4  # https://github.com/Frozenball/pytest-sugar

# Documentation
# ------------------------------------------------------------------------------
sphinx==3.4.3  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2020.9.1 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8==3.8.4  # https://github.com/PyCQA/flake8
flake8-isort==4.0.0  # https://github.com/gforcada/flake8-isort
coverage==5.4  # https://github.com/nedbat/coveragepy
black==20.8b1  # https://github.com/ambv/black
pylint-django==2.4.2  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pre-commit==2.9.3  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.2.0  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==3.2  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.1.0  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.8.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.1.0  # https://github.com/pytest-dev/pytest-django

django-heroku